feat(#457): refactor scripts and document usage
