+++ packages/storage-evm/project.json
@@ -5,10 +5,17 @@
   "projectType": "library",
   "tags": [],
   "targets": {
-    "run": {
+    "script": {
       "executor": "nx:run-commands",
       "options": {
         "command": "npx hardhat run",
+        "cwd": "packages/storage-evm"
+      }
+    },
+    "task": {
+      "executor": "nx:run-commands",
+      "options": {
+        "command": "npx hardhat",
         "cwd": "packages/storage-evm"
       }
     },
