+++ packages/storage-evm/src/index.ts
@@ -1,1 +1,2 @@
+export * from './dao-space-factory';
 export * from './generated';
